refactor: move chart-config into src/ui with re-export shim
